% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdh_data.R
\docType{data}
\name{qdh_spp_mat}
\alias{qdh_spp_mat}
\title{Species presence matrix in Qiandaohu}
\format{
An matrix with row as island and column as species
}
\source{
\href{https://besjournals.onlinelibrary.wiley.com/doi/10.1111/1365-2656.12650}{Si et al. 2017}
}
\usage{
data(qdh_spp_mat)
}
\description{
The species presence in qiandaohu surveyed by Si et al. around 2012.
}
\references{
Si, Xingfeng, et al. "Functional and phylogenetic structure of island bird communities." Journal of Animal Ecology 86.3 (2017): 532-542.
}
\keyword{datasets}
