% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBTpermutation.R
\name{rho_permutation}
\alias{rho_permutation}
\title{test for non-neutral factor using rho statistic and its permutation distribution
rho statistics is the least square coefficient between richness and average trait distance change on an island from grand mean distance (when the island has more than one species), this routine also draw a permutation distribution by switching species labels to test for competition (more positive) and environmental sorting (more negative)}
\usage{
rho_permutation(spp_mat, trait_dist, n = min(2000, gamma(ncol(spp_mat) + 1)))
}
\arguments{
\item{spp_mat}{presence matrix, row as islands/sites, column as species}

\item{trait_dist}{trait distance matrix between species, order should be the same as spp_mat's column}

\item{n}{number of random permutations to draw}
}
\value{
a list with the observed rho, permutation samples and one side/double side p-value
}
\description{
test for non-neutral factor using rho statistic and its permutation distribution
rho statistics is the least square coefficient between richness and average trait distance change on an island from grand mean distance (when the island has more than one species), this routine also draw a permutation distribution by switching species labels to test for competition (more positive) and environmental sorting (more negative)
}
