% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBTpermutation.R
\name{simple_IBT}
\alias{simple_IBT}
\title{simulating a simple IBT presence matrix
simulating a species presence matrix using logistic presence probability depends on distance and size of island}
\usage{
simple_IBT(spp_list, island_feature, betas)
}
\arguments{
\item{spp_list}{string vector of species}

\item{island_feature}{a matrix with three columns, name of island, size of island, and distance to mainland}

\item{betas}{the coefficients for the logistics presence probability, intercept, size and distance}
}
\value{
a species presence matrix simulated using IBT, row as islands
}
\description{
simulating a simple IBT presence matrix
simulating a species presence matrix using logistic presence probability depends on distance and size of island
}
