% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBTpermutation.R
\name{simple_sortingIBT}
\alias{simple_sortingIBT}
\title{simulating a simple IBT presence matrix with simple environment sorting
simulating a species presence matrix using logistic presence probability depends on distance and size of island, while a island feature and a species trait, when they are close the species are more likely to be present}
\usage{
simple_sortingIBT(spp_traits, island_feature, betas)
}
\arguments{
\item{spp_traits}{trait of species, should be a matrix, first column should be name and second being the trait being selected}

\item{island_feature}{a matrix with at least four columns, name of island, size of island, and distance to mainland, the last one is the target trait of an island, the distance between the required trait and the trait of the species determined the chance (together with size and distance)}

\item{betas}{the coefficients for the logistics presence probability, intercept, size, distance and island requirement difference to trait of species}
}
\value{
a species presence matrix simulated using IBT, row as islands
}
\description{
simulating a simple IBT presence matrix with simple environment sorting
simulating a species presence matrix using logistic presence probability depends on distance and size of island, while a island feature and a species trait, when they are close the species are more likely to be present
}
